Group Bodies by Shape
Previouly bodies where only grouped, if they where the
same instance. This means if we drew the same piece
twice, they where treated as individual pieces.

Now the algorithm tries to find existing bodies, where
the shape is equal to the new body. If it finds one, it
increases the count of the existing one, instead of adding
a new one.