Use persistend Naming for plan views
This change adds a Planname property to each body.
This name can then be printed on all pages where the body
is added.

The command Name_Bodies will calculate a name for each
body and set the Planname property. The name consist of
a single letter from A-Z. If more than 26 bodies are found,
it repeats by using AA-ZZ and so forth. Multiple Bodies can
have the same name, if they belong to the same root
object, or their shapes are equal.

The Create_BOM command uses this names and prints it
next to the body count. This makes it easy to reference the
correct element later on in the cut list and manuals.